Optimizing underground duct banks heat dissipation
Heat Jax forward method is used here. By altering the permeability of the backfill and adjusting the spacing between the cables, a reduction in the average temperature within the backfill material is anticipated. As depicted in the figure below, when the permeability shifts from 1E-15 to 1E-8, the temperature decreases from 14 degrees to 4 degrees.